@@ -29,6 +29,8 @@
     "@types/yup": "^0.29.0",
     "chart.js": "^3.6.2",
     "classnames": "^2.2.6",
+    "connected-react-router": "^6.9.3",
+    "cowsay": "^1.5.0",
     "date-fns": "^2.16.1",
     "env-cmd": "^10.1.0",
     "filereader": "^0.10.3",
@@ -46,7 +48,9 @@
     "moment": "^2.29.4",
     "node-fetch": "^2.6.1",
     "numbro": "^2.3.5",
+    "ra-data-json-server": "^4.2.5",
     "react": "^16.13.1",
+    "react-admin": "^4.2.5",
     "react-chartjs-2": "^4.0.0",
     "react-dom": "^16.13.1",
     "react-draggable": "^4.4.4",
@@ -54,18 +58,20 @@
     "react-image-crop": "^9.1.1",
     "react-number-format": "^4.8.0",
     "react-redux": "^7.2.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-scripts": "^5.0.1",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "react-webcam": "^6.0.1",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
+    "redux-saga": "^1.1.3",
     "redux-thunk": "^2.3.0",
     "rifm": "^0.12.0",
     "sass": "^1.26.5",
     "sass-loader": "^10.0.2",
     "susy": "^2.2.14",
-    "typescript": "^3.7.5",
+    "typescript": "^4.7.4",
     "yup": "^0.29.0"
   },
   "scripts": {
