@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { useSelector } from "react-redux";
-import { useHistory, useLocation } from "react-router";
+import { useNavigate, useLocation } from "react-router";
 import { TAPIResponseStatus } from "../../../state/types";
 import { IState } from "../../../types";
 import { IRedirectAfterLogin } from "./types";
@@ -9,18 +9,18 @@ export const RedirectAfterLogin: React.FC<IRedirectAfterLogin> = ({
   children,
   successRoute,
 }) => {
-  const history = useHistory();
-  const location = useLocation<{ from: Location }>();
+  const navigate = useNavigate();
+  const location = useLocation();
   const status = useSelector<IState, TAPIResponseStatus>(
     (state) => state.main.authentication.status || "IDLE"
   );
 
   useEffect(() => {
     if (status === "SUCCESS") {
-      const from = location.state?.from || { pathname: successRoute };
-      history.replace(from);
+      const from = location.state ? location.pathname : { pathname: successRoute };
+      navigate(from, { replace: true })
     }
-  }, [status, location.state, history, successRoute]);
+  }, [status, location.state, navigate, successRoute]);
 
   return <>{children}</>;
 };
