@@ -11,7 +11,7 @@ import Arrow from "../../../assets/arrow-w.svg";
 import { IUserSection } from "./types";
 import "./styles.scss";
 import { useTranslation } from "react-i18next";
-import { useRouteMatch, useHistory } from "react-router";
+import { useLocation, useNavigate } from "react-router";
 
 interface IOwnProps {
   setUserSection: React.Dispatch<React.SetStateAction<IUserSection>>;
@@ -28,15 +28,15 @@ const OutPatientDashboardMenu: FunctionComponent<IOwnProps> = ({
     return value === userSection ? "active" : "default";
   };
 
-  const { url } = useRouteMatch();
-  const history = useHistory();
+  const { pathname } = useLocation();
+  const navigate = useNavigate();
 
   const changeUserSection = useCallback(
     (section: IUserSection) => {
       setUserSection(section);
-      history.replace(`${url}/${section}`);
+      navigate(`${pathname}/${section}`, { replace: true })
     },
-    [history, url, setUserSection]
+    [navigate, pathname, setUserSection]
   );
 
   return (
