@@ -1,3 +1,3 @@
 import { createBrowserHistory } from "history";
-const history = createBrowserHistory({ forceRefresh: true });
+const history = createBrowserHistory();
 export default history;
